requirements.txt
aiofile==3.9.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==26.1.0
Authlib==1.6.9
beartype==0.22.9
cachetools==7.0.5
caio==0.9.25
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
cryptography==46.0.5
cyclopts==4.10.0
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
docutils==0.22.4
email-validator==2.3.0
exceptiongroup==1.3.1
fastapi==0.135.1
fastmcp==3.1.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
importlib_metadata==8.7.1
jaraco.classes==3.4.0
jaraco.context==6.1.2
jaraco.functools==4.4.0
jiter==0.13.0
jsonpatch==1.33
jsonpath-ng==1.8.0
jsonpointer==3.1.0
jsonref==1.1.0
jsonschema==4.26.0
jsonschema-path==0.4.5
jsonschema-specifications==2025.9.1
keyring==25.7.0
langchain==1.2.13
langchain-core==1.2.20
langchain-mcp-adapters==0.2.2
langchain-openai==1.1.11
langgraph==1.1.3
langgraph-checkpoint==4.0.1
langgraph-checkpoint-redis==0.4.0
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.12
langsmith==0.7.22
markdown-it-py==4.0.0
mcp==1.26.0
mdurl==0.1.2
ml_dtypes==0.5.4
more-itertools==10.8.0
numpy==2.4.4
openai==2.29.0
openapi-pydantic==0.5.1
opentelemetry-api==1.40.0
orjson==3.11.7
ormsgpack==1.12.2
packaging==26.0
pathable==0.5.0
platformdirs==4.9.4
py-key-value-aio==0.4.4
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.12.1
pyperclip==1.11.0
python-dotenv==1.2.2
python-multipart==0.0.22
python-ulid==3.1.0
PyYAML==6.0.3
redis==7.4.0
redisvl==0.16.0
referencing==0.37.0
regex==2026.2.28
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.3.3
rich-rst==1.3.2
rpds-py==0.30.0
sniffio==1.3.1
sse-starlette==3.3.3
starlette==0.52.1
tenacity==9.1.4
tiktoken==0.12.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uncalled-for==0.2.0
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.42.0
watchfiles==1.1.1
websockets==16.0
xxhash==3.6.0
zipp==3.23.0
zstandard==0.25.0
